feat: rename obys to observability
