+++ pkg/sidecars/observability/otel.go
@@ -1,4 +1,4 @@
-package obsy
+package observability
 
 import (
 	"fmt"
